Partial Code For SetBasedSearch
While not complete and debugged, this is my current code.

Ext.java contains the extension rules as static methods. For efficiency later on this might have to be changed.
QuickSort.java is just quick sort for floats
QuickSortWithIndex.java is quicksort for floats but also returns an array of what index was moved where
Selection.java Contains f_select, both as a static method, and as an object, which is better for multiple calls on the same population.
SetBasedSearch.java is a very partial implementation of the search control. WIP

This still needs to be debugged, so I have not added these to the project file.