Added a basic "FindTightestBound" Method
added a static getTightestBoundClass method that returns a sorted index array for the classes, where the higher the index of the class, the lower the constraints are. Use this to prioritize classes that are at the front of the array.

Class array structure assumed to be courses then labs, in the the order that they were parsed.